# ENH: improve the speed of numpy.where using a branchless code