--- a/doc/release/upcoming_changes/21130.performance.rst
+++ b/doc/release/upcoming_changes/21130.performance.rst
@@ -1,0 +1,4 @@
+Faster ``np.where``
+-------------------
+`numpy.where` is now much faster than previously on unpredictable/random
+input data.