<!--         ----------------------------------------------------------------
                MAKE SURE YOUR PR GETS THE ATTENTION IT DESERVES!
                ----------------------------------------------------------------

*  FORMAT IT RIGHT:
      http://www.numpy.org/devdocs/dev/development_workflow.html#writing-the-commit-message

*  IF IT'S A NEW FEATURE OR API CHANGE, TEST THE WATERS:
      http://www.numpy.org/devdocs/dev/development_workflow.html#get-the-mailing-list-s-opinion

*  HIT ALL THE GUIDELINES:
      https://numpy.org/devdocs/dev/index.html#guidelines

*  WHAT TO DO IF WE HAVEN'T GOTTEN BACK TO YOU:
      http://www.numpy.org/devdocs/dev/development_workflow.html#getting-your-pr-reviewed
-->

Hello,

This PR speed up the `numpy.where` function using a branch-less implementation when the sparsity is >0.1. It is especially efficient on unpredictable input data because the previous implementation based on a conditional was causing many stalls due to branch missprediction. Such a case happens on random input data and often happens on real-world data. The performance of the branch-less implementation can change from one architecture from another. I do not expect it to be slower than the previous one on all modern x86-64 processors. It might be interesting to check if it is sometimes a bit slower on ARM/POWER architectures.

Here is a benchmark (only bench the modified path of the new code since the rest is left unchanged):

```python
size = 1024 * 1024 // 8
a = np.random.rand(size)
rnd_02 = a > 0.2
rnd_03 = a > 0.3
rnd_04 = a > 0.4
rnd_05 = a > 0.5
all_ones = np.ones(size, dtype=bool)
rep_zeros_2 = np.arange(size) % 2 == 0
rep_zeros_4 = np.arange(size) % 4 == 0
rep_zeros_8 = np.arange(size) % 8 == 0
rep_ones_2 = np.arange(size) % 2 > 0
rep_ones_4 = np.arange(size) % 4 > 0
rep_ones_8 = np.arange(size) % 8 > 0

test_cases = [
    rnd_02, rnd_03, rnd_04, rnd_05, 
    all_ones, 
    rep_zeros_2, rep_zeros_4, rep_zeros_8, 
    rep_ones_2, rep_ones_4, rep_ones_8
]

for test_case in test_cases:
    %timeit -n 1000 np.nonzero(test_case)

for test_case in test_cases:
    assert np.allclose(np.nonzero(test_case)[0], np.arange(size)[test_case])
```

and here is the results on my Intel i5-9600KF processor:

```python
----- OLD -----

213 µs ± 1.35 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
295 µs ± 4.35 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
375 µs ± 5.09 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
408 µs ± 731 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
63.7 µs ± 156 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
83.6 µs ± 229 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
78.7 µs ± 1.13 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
70.7 µs ± 577 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
85.2 µs ± 988 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
91.9 µs ± 93.9 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
77.9 µs ± 366 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)

----- NEW -----

65.5 µs ± 696 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
65.7 µs ± 198 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
65.7 µs ± 181 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
65.1 µs ± 138 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
65.5 µs ± 155 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
65.1 µs ± 205 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
64.8 µs ± 271 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
64.3 µs ± 353 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
65.4 µs ± 474 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
65.4 µs ± 437 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
65.7 µs ± 247 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

As you can see, the new timings are nearly always better on this machine. The new code is up to *6.3 times faster*.