@zephyr111 2022-02-28
Thank you for the results and the suggestions. 

The results are quite surprising because a branch-less code should not be impacted by the content of the input array. Thus, the expected run time should be relatively constant like on x86-64. Consequently, I expect GCC to produce an assembly code with conditional jumps despite the code is branch-less. However, this part of the code appear to generate a good code on the three main kind of platform based on [GodBolt results](https://godbolt.org/z/Ej1Won7rG) (using GCC 11.2 with `-O2`). Alternatively, maybe a complex low-level effect is happening POWER processors that I am not aware of. I can try to tune/optimize the code, but the main problem is I do not have access to any POWER processor (nor ARM yet) to check the performances (or at least the produced assembly code).

Regarding airspeed velocity, I am not very familiar with this tool. I my understanding is correct, it is a tool to automate the analysis of the performance of packages like Numpy over time on many platforms. I found [this page](https://pv.github.io/numpy-bench/#/summarylist) showing the results for Numpy but the current platforms are only Intel ones so far.