@seberg 2022-02-28
It would be awesome if you could add a few additional benchmarks (based on the ones you got) here: https://github.com/numpy/numpy/blob/61dec07ba52c0aafe1149e43bb22ffd0cb4b06c9/benchmarks/benchmarks/bench_function_base.py#L282

@tylerjereddy your benchmarks are a bit mixed, but it looks a lot like a general win anyway?  The best thing would probably to expand those `asv` benchmarks I linked, then Tyler can run them locally if we want to be sure that it is a win on POWER as well.