@tylerjereddy 2022-03-01
Yeah, I can run the actual `asv` benchmarks on POWER9 and ARMv8 when ready, just ping me again. Yeah, the results were mixed--probably a net win, but easier to see with `asv` instead of manual inspection of course.