@zephyr111 2022-03-05
I have optimized the implementation (using loop unrolling to improve the instruction level parallelism, especially on architecture with higher-latency instructions like on most ARM architectures). I also tested this on an ARM Graviton2 machine (Amazon EC2 t4g.small with a default Ubuntu 20 LTS system). Results are also pretty good:

```none
-- OLD --

259 µs ± 1.09 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
350 µs ± 555 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
444 µs ± 1.17 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
502 µs ± 779 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
115 µs ± 700 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
193 µs ± 182 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
153 µs ± 68.5 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
133 µs ± 113 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
193 µs ± 179 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
127 µs ± 203 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
115 µs ± 415 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

-- NEW --
124 µs ± 286 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
126 µs ± 1.28 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
127 µs ± 638 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
126 µs ± 1.33 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
126 µs ± 1.26 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
126 µs ± 1.08 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
124 µs ± 491 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
128 µs ± 965 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
127 µs ± 1.52 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
126 µs ± 736 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
128 µs ± 1.4 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

This is almost always faster and up to *4 times faster*. Results are also stable like on an x86-64 PC.

I also updated the benchmark and tested `asv` which is quite interesting and successfully report the biggest differences.

@tylerjereddy Can you run `asv` on POWER9 on this updated branch?