@rossbar 2022-03-07
Just a quick suggestion re: benchmarking - the `asv compare` command can be used to directly compare two commits and the results are presented in a more digestible way. The `runtests.py` script also exposes this functionality via something like: `python runtests.py --bench-compare main faster-where bench_function_base`.

More info available here: https://numpy.org/devdocs/benchmarking.html