@zephyr111 2022-03-13
> I would love to not need that manual unroll there, but considering the speed-up, and that I am not sure that NPY_GCC_OPT_3 will do the trick due to currently being limited to GCC, happy to go ahead with it.

Me too. I tried with a loop using default compiler parameters and GCC was generating a very bad code with it. It seems that using -O3 is sufficient to generate an efficient code similar to the unrolled loop based on [GodBolt results](https://godbolt.org/z/ovcdM4TYd). I will this in the Numpy code.

I also found out that GCC only generate a branchless code since GCC 11. Before this version, the compiler was visibly not able to generate a branchless code. Clang was able to do that since a very long time (at least since Clang 3.0 released 10 years ago). This certainly explain why @tylerjereddy got mitigated results with GCC 9.4 (that was still a bit faster) and why the results indicated that the execution was not branchless. There was open issues submitted in the GCC bug tracker to discuss about this problem in the past (like [this one](https://gcc.gnu.org/bugzilla//show_bug.cgi?id=97734) for example).